- Fix 64-bit Windows build warnings in desc directory - Change Send/Recv to return ssize_t rather than int.
NOTE: I originally had NaClImcSendTypedMessage be responsible
for converting from ssize_t to a proper nacl ABI type. After talking
with Bennet and looking a little harder at the use of this API, I've
decided to push this responsibility up the chain further and make
the IMC functions return ssize_t as well. That way we can delay the
conversion until the syscall wrapper, which is where it makes the
most sense anyway.

Because of this change, I'll also need to make some changes to the
IMC library. However, it's safe to make this change first because the
IMC lib does not yet have a Win64 build.

Review URL: http://codereview.chromium.org/500016

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@1252 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2