Newlib 1.18 does not like straight use of char in isXXX macro (-Wall warning and so error in NaCl). Indeed, C99 standard says the following:
The header <ctype.h> declares several functions useful for classifying and
mapping characters. In all cases the argument is an int, the value of which
shall be representable as an unsigned char or shall equal the value of the
macro EOF. If the argument has any other value, the behavior is undefined.

I'm not sure if simple cast to unsigned char is the best way to cope, but we need
to fix this... somehow.

Review URL: http://codereview.chromium.org/606044

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@1527 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2