Test to exercise the system when the NaCl application uses a high thread count.
This fixes a bug in nc_threads where a malloc -- to obtain memory for
the thread stack -- return value is used unchecked, so when the thread
count is high and the stack size is large (e.g., the current default
of 512K), the pthread_create call would segfault rather than return an
error.

This fixes a bug in nc_threads thread creation failure returned -1 --
without even setting errno.  The API says a positive integer is
returned to indicate an error -- it's what in syscalls would get
stored into errno, but without having to go through the errno macro or
TLS variable.

This also fixes a bug in the service runtme where thread creation
failures returned -NACL_ABI_ENOMEM when the posix docs say that that
is not a permissible return value; we now return -NACL_ABI_EAGAIN
instead.

The thread max for various architectures were experimentally obtained.
For x86-64, the value was raised beyond 8192 with no problem, but the
max value in the header file will stay at 8192 for now.

Review URL: http://codereview.chromium.org/861006

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@1710 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2