This CL was patched to a ToT chrome (sync'd as of 10pm-ish 4/1/10) and successfully loaded NaCl modules (srpc hw) and passed the UrlAsNaClDesc test.  It buffers streaming data into a shared memory region, so that the in-memory RAM footprint for chrome and for the sel_ldr is never too large.  I have not tested the pepper callback; please advise on whether there's an existing test / how.
I have not done performance tests.  I've left the original
StreamBuffer class in the code for now.  I think it's dead code, and
would like to verify that tomorrow and update this CL if it proves to
be the case.  Meanwhile, please ignore it.

- created a gio subclass that wraps the fixed-sized shared memory
gio_shm subclass which provides unbounded sized shared memory by the
usual size doubling allocation mechanism to achieve O(n) copying cost.

- modified plugin to use GioShmUnbounded rather than a linearly
increasing buffering scheme, and to push the usage of the underlying
NaClDesc via DescWrapper through to the LoadModule RPC, so that the
(non-shared) memory footprint remains modest.

Review URL: http://codereview.chromium.org/1577007

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@1974 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2