Mechanism for plugin to select best nexe for its CPU architecture.
The <embed src="..."> attribute is retained with its existing
semantics (this turned out to be convenient), so this change is
backward compatbible---yay.

The new <embed nexes="..."> attribute contains a table, each row of
which contains a pair of a sandbox (e.g. ARM, x86-32) and the URL of
the nexe for that sandbox.

Also in this change:
- tests/multiarch: Selenium-based integration test using prebuilt nexes
  for several architectures.  (Added ARM executable.)
- Fix typo in macro in sel_ldr-inl.h.
- Add extern "C" around nacl_cpuid.h.
- Factor PortablePluginInterface::CheckExecutableVersion methods
  in common, add checks for file length and ELF magic, and issue
  informative messages on all code paths.
- Minor fixes to irritating bug in nacl_js_lib.js
  (which I didn't end up using).
- Fix bug whereby incorrect (local) URL is reported via embed.src
  attribute.
- Issue informative errors to the user when module doesn't appear
  to be an ELF file.

Review URL: http://codereview.chromium.org/1516002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2219 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2