Remove thread refcount and remove NaClAppDtor()
Following r2358, where support for signal stacks was added,
NaClAppThreadDtor() must now be called from the thread that is
shutting down, or the signal stack will be unregistered for the wrong
thread.

This means the refcount for threads should be removed, to ensure that
it is not used unsafely, so remove NaClAppThreadIncRef() (which was
not used anyway) and remove NaClAppThreadDecRef().

NaClAppDtor() called NaClAppThreadDecRef() but is basically unused,
because it is not able to shut down running threads safely.  There is
a TODO for its removal, so remove it.

This leaves other code unused, so remove it too:

 * NaClAppFreeAllMemory()
 * NaClAppFreeWalker()
 * NaClTeardownMprotectGuards()
 * guard_pages_initialized field

web_worker_stub.c called NaClAppDtor() but is dead code.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=560
TEST=existing tests

Review URL: http://codereview.chromium.org/2659002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2458 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2