x86-32 validator: Remove requirement that code chunks end with HLT
An earlier change makes it the responsibility of the loader/runtime to
ensure that the process's code region ends with HLT.

Update tests: To avoid changing every test case and having to change
all the testsize values, we leave the trailing HLTs in, but don't pass
them to the validator.  They still have a use though: acting as a
sanity check on testsize.

Remove test 2: it was the same as test 1, but without a trailing HLT.

Remove test 13: it was the same as test 10, but without a trailing HLT.

Update test 4: The description was not accurate.  It only had
sawfailure=1 before because it lacked a trailing HLT.

Update test 21: Add disassembly.  Add a trailing HLT to match the
sanity check.

TEST=run_dynamic_load_test validator_tests
BUG=http://code.google.com/p/nativeclient/issues/detail?id=585

Review URL: http://codereview.chromium.org/2720006

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2493 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2