Changes to how we build PNaCl untrusted toolchain:
Instead of tarballs+patches, each component lives in its own Mercurial
repository, a sub-repo of nacl-llvm-branches.googlecode.com.

The four components are: newlib, llvm-gcc, binutils and llvm.

 http://code.google.com/p/nacl-llvm-branches/source/list?repo=newlib
 http://code.google.com/p/nacl-llvm-branches/source/list?repo=binutils
 http://code.google.com/p/nacl-llvm-branches/source/list?repo=default
 http://code.google.com/p/nacl-llvm-branches/source/list?repo=llvm-gcc

The script makes a mirror (hg clone) of each repo, saving the mirrors
across runs as an optimisation.  In each build, each mirror is cloned
again (local => fast), and the builds are performed.

For each component, the user may specify the name of a directory in an
environment variable, that causes that dir to be copied instead of the
Hg repo.  This facilitates development builds.

Also:
- Renamed shell functions to be like the corresponding command-line
  actions.
- Used the same dispatch mechanism as pnacl-helper.
- I will put the functions next to their docstrings in a followup
  change; it would make this change inscrutable.

Review URL: http://codereview.chromium.org/2867004

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2515 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2