Landing http://codereview.chromium.org/2821012/show Preparations for valgrind (memcheck) build bot.
We need a command line which will run a specified set of tests
under valgrind (memcheck) on a buildbot.
With this change such command line will be:
./scons --mode=dbg-linux,nacl platform=x86-64 sdl=none buildbot=memcheck memcheck_bot_tests
As of now, this command passes.
It is important to setup the build bot asap so that the tests continue passing under memcheck.
NaCl is a moving target for valgrind and we'd like to make sure we handle NaCl
properly every time it changes.

The patch is prepared by Evgeniy Stepanov.
Review URL: http://codereview.chromium.org/2846023

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2554 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2