fake_browser_test: Extend to run sel_ldr on a real executable
Extend the test to cover getting the plugin to fetch and launch an
executable using sel_ldr.  This involves adding dummy URL-fetching
code and a fake window.location object for the origin check.  We test
invoking an SRPC method.

Move the test to tests/fake_browser.  Now that it relies on building a
NaCl executable it doesn't seem appropriate to keep it in
src/trusted/plugin.

Switch to using dlopen() on the real NPAPI plugin instead of
statically linking the test with the plugin code.

Add NACL_SEL_LDR environment variable to sel_ldr_launcher.cc so that
the plugin can locate sel_ldr on Mac OS X.  (On Linux, the plugin
looks for sel_ldr in the directory the plugin was loaded from, which
works because both are in "staging" in the Scons build.)

This test is able to reproduce
BUG=http://code.google.com/p/nativeclient/issues/detail?id=652
by uncommenting a line.

Review URL: http://codereview.chromium.org/2883007


git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2592 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2