Plugin: Move responsibility for freeing ServiceRuntime object
The plugin does connect() three times to create three connections to
sel_ldr.  The first is to sel_ldr's trusted side, the other two to the
untrusted side.  To the third ConnectedSocket the plugin bestows the
responsibility of freeing its ServiceRuntime object.

In order to simplify things, move the responsibility for freeing
ServiceRuntime to the Plugin object, where it is created.

Aside from this, ConnectedSocket only looked at whether its
ServiceRuntime* argument was NULL or non-NULL, so replace this
argument with a boolean, can_use_proxied_npapi.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=652

Review URL: http://codereview.chromium.org/2878013

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2597 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2