Enabling the new version of newlib, which is at 1.18.0. (Renamed newlib directory from newlib-1.17.0 to newlib-trunk)
newlib has a "pnacl" architecture target now. We are now using that, but as a compromise, renaming the installation directory back to arm-none-linux-gnueabi to avoid breaking everything else.

Added a check for making sure that repos are up to date. utman.sh will prompt if a repo is not at "stable".

Tested with:
  utman.sh clean
  utman.sh all
  utman.sh test-x86-32
  utman.sh test-x86-64
  utman.sh test-arm
  utman.sh test-spec ... SetupPnaclX8664Opt
  utman.sh test-spec ... SetupPnaclX8632Opt
  utman.sh test-spec ... SetupPnaclArmOpt
Review URL: http://codereview.chromium.org/3039023

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2762 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2