Fix handle_pass build breakage
Fix breakage introduced by r2811.  That change removed an argument
from NaClNrdXferEffectorCtor(), but I missed updating handle_pass
because it is only used in Chromium and not tested or built by the
NaCl trybots.

That caused these errors in the Windows Chromium build:

.\ldr_handle.cc(56) : error C2660: 'NaClNrdXferEffectorCtor' : function does
not take 2 arguments
.\browser_handle.cc(121) : error C2660: 'NaClNrdXferEffectorCtor' : function
does not take 2 arguments

Since the socket argument is removed, this code no longer needs to
create a socketpair to pass in.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=344
TEST=build on Windows with http://codereview.chromium.org/3079004 applied

Review URL: http://codereview.chromium.org/3054023

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2815 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2