thread_test flakiness uncovers 3 issues.
Turns out examining this bug uncovered several issues.
1) Too many calls to NaclBreakpadInit - spawns extra threads
2) Race condition in NaClThreadCtor v NaClThreadDtor
3) Assumiption in thread_test.


Currently we call init during module init, plus each NaCl thread created.  
In effect, this is causing an extra thread per and extra chaining of the
vectored exception handler per NEXE thread.  I have removed per thread 
call to breakpad init, and made the global g_pBreakpad truly global
instead of thread local, so that we have one shared copy of breakpad.

Next, I changed _beginthreadex to start the thread suspended.  This allows
us to store the return value (thread handle) in the NaClThread structure
before the thread starts running.  As is, since the Ctor is called by the
parent and the Dtor by the child, there is a race where a short lived
child can exit and call closehandle on itself before the value is ever
stored in the structure.

Finally, there is an assumption in the thread_test that the toolset is
correct.  If for some reason the intrinsic atomic instructions are not
properly handled, we could get into an infinite loop which would produce
a hung test much in the way we have observed.   Unfortunately I have not
been able to reproduce the problem, but I'm updating the test just in case.


Review URL: http://codereview.chromium.org/3022015

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2821 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2