sel_ldr_launcher: Simplify to use ExportImcFD() method internally
Simplify Launch() to use the ExportImcFD() for the socket that is
created by default.  This reduces duplication that was introduced when
ExportImcFD() was added based on the code in Launch().

Make the imc_fd/channel_number_ argument to SelLdrLauncher optional,
such that specifying -1 does not create this socket.

This means we no longer need InitChannelBuf(), $CHAN and $CHAN_BUF.

This means ExpandVar() was only expanding $SEL_LDR_EXE and $APP_EXE,
so it is simpler to expand out the template list into three
push_back() calls so that ExpandVar() is not needed.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=642
TEST=run_fake_browser_test

Review URL: http://codereview.chromium.org/3015040

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2828 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2