Plugin: Hook up async send/receive sockets in Chromium
Add sel_ldr_launcher_chrome.cc to the Scons build so that it gets
compile-tested in the standalone build.  This also reduces the number
of #ifs.
 * Move launch_nacl_process out of nacl_entry_points.cc because that
   cannot be built in the standalone build because of its #includes.
 * Add an extra conditional to the NaClHandlePass call.

Rename the Chromium versions of Start() to StartUnderChromium() to be
more greppable.

Change StartUnderChromium() to use multiple FDs, and change
sel_main_chrome.cc to accept these FDs.

We stop using SelLdrLauncher's channel() method in the plugin, and
pass this value to InitCommunication() as an argument.

The new LaunchNaClProcess type happens to be the same as the old type,
which creates a potential for confusion because the arguments are used
differently.  Remove the argument from the old variant of
RegisterInternalNaClPlugin() to avoid problems.  (This old variant is
still called with no arguments by Chromium in the browser process.)

BUG=http://code.google.com/p/nativeclient/issues/detail?id=642
TEST=nacl_ui_tests in Chromium

Review URL: http://codereview.chromium.org/3044033

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2834 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2