Revert r3037, which breaks the toolchain build
r3037 added a patch from Git that is supposed to be applied to
binutils, but it applied it to gcc, which fails.

Also r3037 was committed using a code review URL that doesn't
correspond to the change.  It looks like it wasn't committed using
gcl/git-cl.  Please reapply the working part with the correct review
URL and the correct format.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/3197004

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@3040 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2