Expanded testing of stores in the ARM validator unit test.
The new test uses lists of valid and invalid instructions, and produces their Cartesian product of programs.  It also modulates the "cond" field in the instruction, testing every combination of conditional instruction.  As a result, this runs 2,580 distinct test programs through the validator.

As you might expect, the runtime has increased.  On an Opteron it takes 13ms.  The same machine running the test in QEMU takes 285ms.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=823
TEST=run_arm_validator_tests

Review URL: http://codereview.chromium.org/3110028

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@3048 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2