Re-enable syscall_return_sandboxing test to work on ARM
All this needed was a data bundle marker.  The test must have
originally passed through fluke, because the .word value happened to
be an allowed instruction.  I imagine it got disabled when the .word
value changed.

The test's use of assembly was not why it got disabled for PNaCl -- in
fact, there is a special case in llvm-fake.py to handle it.

Also address the TODO in nacl.scons.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=777
TEST=run_syscall_return_sandboxing_test

Review URL: http://codereview.chromium.org/3436015

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@3326 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2