DL toolchain script fails on Windows
At some point the download_toolchain.py script had a 'Retry' function added to it which retries certain directory
operations when running under win32 after a short sleep.  Apparently this is to fix a race condition in which
modifications to a directory would fail because outstanding handles were not completely released. Code espections
does not show an obvious issue, since access (untar, move, rm) are serialized.

The retry code was overly restrictive in which expections it would catch.  Originally the funciton would catch
exceptions of type "WindowsError".  However, in this particular case, the error is:
  'OSError: [Errno 13] Permission denied'

This error was not caught and retried.

Changing the function to always retry windows exceptions fixes this issue

In addition, this change reduces the number of retries from 10 to 5, which because of exponential backoff, would
have resulted in over 15 minutes of retrying on an unrecoverable error potentially delaying automated error reporting.
The new timeout should orcur in aproximately 30 seconds.

bug=http://code.google.com/p/nativeclient/issues/detail?id=1130

Review URL: http://codereview.chromium.org/4143014

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@3602 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2