Modular build: Add cache for file contents hashing
There is one cache per build target.  Only the cache results for the
last pass over the build target's contents are kept.  This means that
the cache contents do not accumulate over time.

This makes it faster to rebuild/rescan a target when most of the
contents haven't changed, such as when rebuilding a TreeMapper target
when only one of the inputs has changed.

More importantly, this will make it efficient to scan source
directories on every run in the future.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=484
TEST=python test_all.py

Review URL: http://codereview.chromium.org/4217003

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@3636 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2