Add "executable" argument to CreateMemoryObject()
This enables dynamic loading support in Chromium's Mac sandbox.

This allows the function to distinguish nacl_text.c's calls, since for
this case the memory object needs to be mappable with PROT_EXEC.  We
need to avoid using shm_open() for this on Mac OS X.

We change the Windows implementation to switch between
PAGE_EXECUTE_READWRITE and PAGE_READWRITE for completeness, and as a
possible defence in depth, although it is not immediately required.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=583
TEST=test_runner.html with Chromium-side change applied, on a Mac

Review URL: http://codereview.chromium.org/4695002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@3664 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2