Add sched_yield() calls to simple_thread_test.
This tests uses a mutex to transfer control between threads in a ping-pong fashion.
Since by default pthread mutex is not fair, this can cause starvation in one of the threads.
This is likely the reason of the recent timeouts on the Valgrind bot.
Review URL: http://codereview.chromium.org/5178004

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@3721 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2