Python IMC: Add support for setting up a connection to a subprocess
This involves adding a couple of functions that operate on raw host-OS
sockets rather than on NaClDescs.

In order to get the test subprocess to run with x86-32 Python on Mac
OS X 10.6, I had to pass an environment variable through.  Supposedly
the OS-provided Python provides "VERSIONER_PYTHON_PREFER_32_BIT=yes"
for doing this, but it did not work on the machine I tested on.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1230
TEST=run_python_imc_test

Review URL: http://codereview.chromium.org/6199003

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@4099 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2