PPAPI Plugin: Fix proxying of the scriptable instance object.
Before this fix, the trusted interface was returned if the nexe and proxy setup was
still in progress and a separate proxied interface was returned otherwise. Thus, 
if the browser asked for the scriptable object too early, it would be stuck scripting
the NaCl plugin. The fix makes the proxied interface part of ScriptableHandlePpapi, 
set when proxied execution is started, so NaCl can properly redirect scripting
requests to the proxy, once it is ready.

Also, address related TODOs for shutting down the proxy as well as error checking and caching
PPP_Instance interface pointer.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=1325
TESTED=manually with examples_ppapi.html

Review URL: http://codereview.chromium.org/6310014

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@4193 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2