Exit Cleanup
Add a nacl_exit.h and nacl_exit.c to the src/shared/platform library. Convert all instances of 'abort' and '_exit' to nacl_abort and nacl_exit within the trusted code base (not including tests). This allows us to use a fast exit when an error is detected.  This CL also looks for the macro COVERAGE signalling a coverage build, to bypass the quick exit and use a normal exit to allow coverage information to be flushed out before exiting.
While there is an assumption that _exit or abort are used in all cases where fast exit is required, the list of changes was derived by manual verification of a grep for all instances of abort and exit (instead of abort( and _exit( since the may be used as a function pointer, or exit may be used when _exit was really needed).   We run all tests to verify all tests that require specific exit codes run correctly.

In addition to changes of _exit and abort, this CL also modifies our use of sysv_shm_and_mmap.c during platform qualification to avoid the use of atexit.

TEST=all
BUG=nacl1561
Review URL: http://codereview.chromium.org/6676106

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@4767 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2