IRT: Implement transferring control to user executable at startup
Change sel_ldr to fill out AT_ENTRY in the auxv when using "-B":
 * Rename NaClApp's entry_pt to initial_entry_pt.
 * Add user_entry_pt for passing via AT_ENTRY.
 * Add a CHECK() to ensure that ptr_tbl_size matches up with what was
   written via BLAT().  Note that previously, ptr_tbl_size was actually
   8 bytes larger than necessary.

Add IRT code for jumping to the user executable.  This needs some
assembly code for invoking the executable's ELF entry point.

Move src/untrusted/irt/nacl.scons to be built from --mode=nacl rather
than --mode=nacl_extra_sdk.  This is more correct (the IRT doesn't
belong in the toolchain tarball) and more convenient (just one Scons
invocation is needed to build and test).  This requires removing
"AddLibraryToSdk(['libirt'])", but tests/irt_hw will still find the
library OK.

Add "irt" option to Scons so that we run the Scons tests with the IRT
enabled.  A few don't pass initially and this is not the default on
the buildbots.

Move address space layout settings to SConstruct so that they are
shared.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1595
TEST="run_blob_library_loading_test" plus "run_hello_world_test irt=1"

Review URL: http://codereview.chromium.org/6800027

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@4782 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2