Debug stub: Refactor startup to use a single function to enable debugging
This should give us more confidence that, if the debug stub is later
compiled in by default, its code is not called accidentally.  It would
also make it easier to build debug and non-debug versions of sel_ldr.

Remove service_runtime's direct calls to NaClDebug*() functions and
replace them with callbacks that may be registered by the debug stub.

Remove the "-b" option since it did not do anything (break-on-start
was always enabled).

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1656
TEST=run_debug_stub_test with "#define NACL_DEBUG_STUB" uncommented

Review URL: http://codereview.chromium.org/6839026

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@4876 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2