Browser tests: Use xvfb-run instead of VNC on Linux
The current code for launching a dummy X server hardcodes two values:
 * The pathname /home/chrome-bot/.Xauthority
 * The display number :20
and it also attempts to shut down the X server.

However, that is unreliable.  After a recent restart of the bots, some
are failing with the output:

    + vncserver -kill :20
    Can't find file /home/chrome-bot/.vnc/nacltest1:20.pid
    You'll have to kill the Xvnc4 process manually
    + true
    + sleep 2
    + vncserver :20 -geometry 1500x1000 -depth 24
    Warning: nacltest1:20 is taken because of /tmp/.X20-lock
    Remove this file if there is no X server nacltest1:20
    A VNC server is already running as :20

Use Xvfb instead, since the xvfb-run script will allocate an unused
display numbers and take care of shutting down the X server.  The
xvfb-run script is not perfect, but it will probably be more reliable
than our vncserver invocation.

BUG=none
TEST=trybots

Review URL: http://codereview.chromium.org/6737038

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@4921 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2