IRT: Archive irt.nexe from the buildbots for use in Chromium
Upload the integrated runtime (IRT) binary so that Chromium can pull
it in via DEPS, and so that the Chromium build does not have to be
able to build NaCl-targetted code.

The upload happens from the main buildbots rather than the toolchain
buildbots because the IRT is not part of the toolchain.

This is a retry of r4975 with some fixes:
 * Move the upload step so that it's not run on the glibc or toolchain bots.
 * Fix the STEP_LINK text:  add missing "@@@".
 * Use "lucid32-m32-n32-dbg" rather than the "-opt" version, because the
   "-dbg" build slaves run sooner -- they are less overloaded.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1595
TEST=I tested modified versions of this on the trybots

Review URL: http://codereview.chromium.org/6895007

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@4980 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2