Fix debugger stub not to skip first instruction byte on x86
This works around issue 1730 but is not really a proper fix for it.

Firstly, we modify the debug_stub_test so that it notices when its
argc/argv are completely bogus, which is the status quo.

Second, we change the debug stub to use HLT rather than INT3 as the
breakpoint instruction.  The stub directly choosing a breakpoint is only
used in the case of the breakpoint inserted at the entry point to catch
the program at startup.  For this, nothing really cares what kind of
exception is cited, so it being a general protection fault exception
rather than a breakpoint trap exception doesn't matter.

A better fix would be to make the stub recognize when a breakpoint it
has inserted itself has been hit, and back the PC up by one byte.
There is some concern that perhaps the Windows trap handler does this
adjustment itself, so that needs to be looked into.  If it's the case
that it does, then this is probably also a problem for GDB talking to
a Windows-based debug stub, since GDB will not expect the PC to be
adjusted this way (it's not what the hardware itself does).

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1730
TEST=modified debug_stub test fails without workaround, works with it

R=mseaborn@chromium.org,mlinck@google.com

Review URL: http://codereview.chromium.org/6883141

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@5021 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2