This defines what might be the stable IRT ABI in untrusted/irt/irt.h, implements it in the IRT, and makes the newlib build use it for system services.
This is still very much a transitional state, far from clean or final.
Right now, the same newlib/libnacl arrangement is used to build
irt.nexe itself, the miscellaneous private tests, and newlib-based
user applications.  For this to work, the newlib/libnacl code has to
work both with and without an IRT.  It does this by always using
indirection tables for the calls that should go through the IRT, but
falling back to supplying its own tables when there is no IRT present.
These are the same tables that are built into the IRT itself to
provide the official ABI entry points.

Later, we should split the library providing the system interface
calls via the IRT from the one providing private interfaces used
in tests and in the IRT itself.  Doing that requires cleaning up
the linking rules used by the buildsystem and toolchain.

The irt_hw test is removed because it no longer makes sense in the
context of true IRT loading.  All normal tests now constitute IRT
tests when scons it given the irt=1 parameter.

R=mseaborn@chromium.org,bradchen@google.com
BUG=http://code.google.com/p/nativeclient/issues/detail?id=1595
TEST=chrome_browser_tests irt=1 works;
     {small,medium,large}_tests irt=1 works as well as they did before;
     trybots

Review URL: http://codereview.chromium.org/6899028

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@5108 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2