Install -lppapi that includes ppapi_stub library and places .rodata after gap
This installs a tiny linker script under the name libppapi.a for use
with linking programs using PPAPI via the IRT (as -lppapi).  This brings
in the libppapi_stub.a library, and also defines a magic symbol that
makes the NaCl linker scripts place R/O data after a large gap, leaving
space for the IRT code.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1595
TEST=trybots

R=sehr@google.com,bradchen@google.com,mseaborn@google.com

Review URL: http://codereview.chromium.org/6927067

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@5208 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2