Update pyauto NaCl tests to use python 2.6 on Linux
The pyauto NaCl tests required python 2.5 because the pyauto libs they
depend on were pre-built with python 2.5 on Linux.

With r86985, pyauto is built with python 2.6 on linux, as a result of
which the pyauto nacl tests no longer require python 2.5. In addition,
the location of _pyautolib.so on the chrome build webpage has changed.

This patch fixes SConstruct to use python 2.6 for pyauto NaCl tests on
Linux. It also fixes the chrome download scripts to use the new location
of _pyautolib.so. Finally, it bumps the chrome_rev in NaCl to r87002.

BUG=nacl:1571
TEST=scons pyauto_tests

Review URL: http://codereview.chromium.org/7086011

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@5492 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2