sel_ldr: Use single VirtualAlloc() calls to map guard regions on 64-bit Windows
This will make startup faster.

Previously, each 64k page of the 40GB guard regions was mapped with a
separate VirtualAlloc() call.  This resulted in 1310720 VirtualAlloc()
calls, which caused address space allocation to take around 1 second.

We can fix this by doing just two VirtualAlloc() calls, because the
pages in the guard regions do not need to be independently remapped
(unlike the main 4GB part of untrusted address space).

We create sel_addrspace_win_x86_64.c with the code that is specific to
both x86-64 and Windows.

We create sel_addrspace_posix_x86_64.c and move some functions here:
 * Move the POSIX version of NaClAllocatePow2AlignedMemory() here.
   We no longer compile this function in on x86-32, where it was not
   used.  The Windows version is no longer needed.
 * Move the original x86-64 implementation of NaClAllocateSpace() here.
   This is now used only on POSIX.

Extend mmap_test.c to check that the guard regions are correctly
mapped on Windows.  Give this test a name in Scons so that it can be
run directly.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1876
TEST=run_trusted_mmap_test

Review URL: http://codereview.chromium.org/6982044

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@5549 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2