Make optimization controllable via pnacl driver
In particular, do not run opt at link time unless -On (where n != 0) is specified. Previously opt would be run at link time unconditionally.

For the user this just means that optimization will happen if and only if they ask for it, both at compile and link time.

Internally this changes the OPT_LEVEL env var to just the number (or 's') instead of the whole '-On' string

TEST=scons, selfbuild

Review URL: http://codereview.chromium.org/7044085

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@5709 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2