ARM: Use host TLS instead of r9 low bits for tracking current NaCl thread
We can use the host TLS (i.e. a simple __thread variable) to keep track
of which NaClAppThread a particular host thread is.  This relieves the
untrusted ABI of the requirement to preserve some low bits in r9 for
trusted code's use, so the r9 value (untrusted thread pointer) need not
be aligned to something huge.

Getting rid of the artificial alignment requirement saves the untrusted
code from having to waste so much memory for alignment padding for each
thread.

This also removes the fixed limit on the number of threads.  We still
have an arbitrary limit (now 8192 rather than 4096), but this is purely
an implementation detail in the service runtime.

Note that with this there is no security/safety reason that the sandbox
has to disallow untrusted code from changing r9.  But reserving r9 for
the thread pointer in the untrusted ABI is still a good thing, and the
only benefit of untrusted code setting it directly would be to optimize
nacl_tls_init from being a trampoline/syscall to being a simple register
move.  Since on x86 nacl_tls_init must be a syscall anyway, there is no
reason to bother with that.

BUG=cleanup
TEST=trybots

R=bsy@google.com

Review URL: http://codereview.chromium.org/7190043

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@5763 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2