fix a race condition in reverse channel startup.
reverse channel initialization state is now a 3 state state variable
rather than a binary flag.  after reverse_setup RPC to send the
connection capability to the plugin is invoked, it goes into an
in-progress state, and when the first reverse RPC connection is made
the reverse channel initialization state becomes fully initialized.
the service thread spawner can detect when initialization is
in-progress and wait for the reverse channel is fully initialized
before constructing the manifest proxy object.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=1949
TEST= None
Review URL: http://codereview.chromium.org/7240020

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@5791 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2