x86-64 Mac: Use the 32-bit Chromium binary for platform=x86-64
Use the same hack as with Windows for the mixed 32/64-bit case.

This allows chrome_browser_tests to be run for x86-64 Mac, but none of
it passes yet: sandbox_isa.cc needs to be changed so that
GetSandboxISA() can return "x86-64".  Furthermore there is another
problem in which the communication between the plugin and sel_ldr does
not complete, and they hang during startup.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1979
TEST="run_ppapi_ppb_core_browser_test platform=x86-64" on Mac

Review URL: http://codereview.chromium.org/7316003

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@5936 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2