x86-64 sandbox: Avoid revealing NaClThreadContext pointer to untrusted code
If there were a vulnerability that allowed writing to an arbitrary
address, knowing the address of struct NaClThreadContext would make
the vulnerability easy to exploit in a reliable, cross-platform way.
So to be on the safe side, we avoid leaking this address.  The cost is
only the same as the cost untrusted code already pays for all indirect
jumps.

This also removes a source of non-determinism.  Others remain, of course!

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2014
TEST=no specific test for the new behaviour

Review URL: http://codereview.chromium.org/7375004

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@6059 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2