Move egyptian_cotton_test from small_tests to medium_tests
This test usually makes my Linux laptop thrash and renders it unusable
for some time when I run small_tests.

I'm not moving egyptian_cotton to large_tests, on the grounds that
large_tests is supposed to be for tests that do not safely run
concurrently.

BUG=none
TEST=small_tests works locally

Review URL: http://codereview.chromium.org/7129049

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@6091 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2