On ARM in QEMU, pthread_cond_broadcast appears to wake up only one thread instead of all of them. In order to work around this, we have a hack in NaClCondVarWait which always uses a timedwait on ARM.
This wastes a lot of CPU time in spinning: acquiring and releasing
mutexes when the condition times out but hasn't been signalled yet.

Remove this hack and instead keep a counter which counts how
many threads are waiting for the signal. In NaClCondVarBroadcast,
use the counter to determine how many times to repeatedly
send a signal.

Before this change, egyptian cotton with 1000 threads (in QEMU) takes 55 seconds.
After this change, it takes 14 seconds.
Review URL: http://codereview.chromium.org/7278029

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@6142 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2