Add "bigalloc" test (malloc of 128M)
Add a new test that's a trivial program allocating a 128M block with
malloc.  This test currently works as is, but fails under the IRT.
(Filing an issue related to that shortly.  It looks like a service
runtime bug.)  The test will probably always work under glibc.

BUG=none
TEST=trybots

R=sehr@google.com,bsy@google.com

Review URL: http://codereview.chromium.org/7472043

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@6155 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2