Ensure that the exit syscall really exits the process in Chromium
Change NaClMainForChromium() to exit the process.  It was assuming the
caller on the Chromium side would exit, but the caller was not
exiting.

Change the function's return type to be consistent with this.  Using
'void' is fine since the Chromium side does not look at the return
value anyway.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2083
TEST=run_ppapi_crash_browser_test currently triggers this because its
  CHECK() call happens to do an _exit()

Review URL: http://codereview.chromium.org/7503005

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@6200 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2