Ensure nacl_thread[] entry is not left set if thread creation fails
Previously, if the creation of the host OS thread failed, the entries
in the nacl_thread/nacl_user/nacl_sys arrays would be left behind,
pointing to a dangling NaClAppThread.  This does not matter too much,
because no-one should dereference those entries, but if
NaClSysCommonThreadSuicide() is going to go to the trouble of cleaning
out these entries, we might as well ensure it happens in all cases.

This change also helps make thread setup and teardown more
symmetrical, so that:
 * NaClThreadLauncher()'s setup is undone in NaClSysCommonThreadSuicide()
 * NaClAppThreadCtor()'s setup is undone in NaClAppThreadDtor()

BUG=none
TEST=trybots


Review URL: http://codereview.chromium.org/7583022

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@6429 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2