x86-32 sandbox: Factor out duplicated code in nacl_syscall_32.S
Share the code that is common between the PIC and non-PIC cases.

This is a retry of r6357, which contained a mistake.  Previously I
only did a visual diff of the code, and I missed the nacl_sys PIC
case, and this broke at link time in the Mac Chromium build.  This
time there is a test target to catch this (dummy_sel_ldr.dylib).  I
have been more careful by running the two #if branches through "diff",
which shows the following differences:

 .globl  IDENTIFIER(NaClSyscallSeg)
         HIDDEN(NaClSyscallSeg)
 IDENTIFIER(NaClSyscallSeg):
         cld
         xor     %eax, %eax
-        movw    %gs, %ax
+        mov     %gs, %ax
         shr     $3, %eax
-        movl    0(%ecx), %edx
-        movl    0(%edx,%eax,4), %edx
+
+        movl    IDENTIFIER(nacl_user)(,%eax,4), %edx
 
         /* check NaClThreadContext in sel_rt_32.h for the offsets */
         movl    %ebx, 0x0(%edx)
@@ -87,8 +71,7 @@
          * any registers having particular values.  we will clear/discard
          * caller-saved registers at system call return.
          */
-        movl    4(%ecx), %edx
-        movl    0(%edx,%eax,4), %edx
+        movl    IDENTIFIER(nacl_sys)(,%eax,4), %edx
         /*
          * %cs and %ds already taken care of by NaCl_trampoline_seg_code
          *

BUG=none
TEST=run_hello_world_test on Linux (currently non-PIC) and Mac (PIC)
  plus the dummy_sel_ldr.dylib target on Mac


Review URL: http://codereview.chromium.org/7640026

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@6432 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2