Start low memory reservation at 0, not 4k
It is entirely possible to have mmap_min_addr set to zero.
If a system does that, we need to protect the bottom page.

Another note: it seems unwise here to use NACL_PAGESIZE, rather than
dynamically getting the system page size in the usual fashion.  Our
setting is 4k and that is indeed the x86-32 page size and almost
certainly always will be.  But the macro's name and definition connote
something we've chosen for NaCl, which might be larger than the system
minimum page size.

BUG= none
TEST= trybots

R=bsy@google.com,bradchen@google.com

Review URL: http://codereview.chromium.org/7778047

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@6600 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2