Make command_tester.py's output look more like gtest's.
For example, "[  FAILED  ] testname (X ms)" is printed out when a test fails.
The allows buildbot hooks / Chrome extensions that are designed to deal with
Chrome's tests also work for NaCl's.  SCons has also been modified to print
test names as the alias that can be used to run the test rather than the
bogus filename used to represent a test in the dependancy graph.  This should
make it easier for sheriffs to run and disable failing tests.

BUG= http://code.google.com/p/chromium/issues/detail?id=95288
TEST= none

Review URL: http://codereview.chromium.org/7848006

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@6676 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2