Always use proper PIC compilation for shared objects on Linux (Gyp)
Building non-PIC code into shared objects is never a good idea.
It works on x86-32, but is highly discouraged.  On some systems,
it requires that Chromium be marked to have special extra privilege.
It only works by happenstance on ARM, and starts to fail when some
slightly different code generation comes along.

Note: this only affects the standalone Gyp build.  Corresponding
changes for the SCons build are already in flight.  The build of
the NaCl code in Chromium is already done as all proper PIC.

BUG= none
TEST= trybots

R=bsy@google.com,bradnelson@google.com

Review URL: http://codereview.chromium.org/8046009

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@6786 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2