Disable some browser tests under Valgrind.
Part of them are disabled because they make it difficult to detect
the nexe path, or use multiple nexes. Valgrind can not handle that at
the moment. Such tests are marked with TODOs.

The rest are crash tests. We don't want to run them under Valgrind.
Review URL: http://codereview.chromium.org/8359031

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@7003 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2