Major cleanup to utman.sh
Introduce FRONTEND environmental parameter which controls the build frontend.

The allowed values for FRONTEND are:

default: Use whatever frontend is known to work. (This is Clang for everthing except libstdcpp and self-build). This is the same as omitting FRONTEND.

clang: Use pnacl-clang for everything

llvm-gcc: Use pnacl-gcc for everything

dragonegg: Use pnacl-dgcc for everything
Review URL: http://codereview.chromium.org/8373020

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@7014 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2