command_tester.py: Add new option for testing exit statuses for crash tests
Allow test programs to declare their intended exit status by printing
it to stderr with a line of the form "** intended_exit_status=X".
With the option --declares_exit_status=1, command_tester.py will
require such a line in stderr and will check it against the actual
exit status.

This allows crash tests to be less coupled to Scons because the exit
status won't have to be declared in Scons.  It also removes the need
for crash tests to have golden files.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2442
TEST=I've converted run_execute_data_test

Review URL: http://codereview.chromium.org/8565031

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@7344 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2