Don't misalign the trusted stack on ARM
We can't figure out why this + 4 was there in the first place.  It seems
always to have been wrong.  It only matters in practice when a confluence
of circumstances winds up with something in the trusted code depending on
the normal ABI stack alignment, so it has gone unnoticed for a long time.

BUG= none
TEST= trybots

R=bsy@google.com

Review URL: http://codereview.chromium.org/8790007

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@7355 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2