Syscall list: Use NaClCommonSysImc_SocketPair() directly
Stop going via the OS-specific files for this syscall.

This is one special case I'm addressing before doing this for the
other syscalls.

In order for the function's type to match the type listed in
SYSCALL_LIST, we remove the "volatile" qualifier, which none of the
other syscall functions use for pointers in their type signatures.

Furthermore, it is more accurate to use "uint32_t" for the function's
argument because this is an untrusted pointer.  This is cleaner
because it avoids the need to cast the argument to
NaClUserToSysAddrRange().

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2540
TEST='run_simple_srpc_test_nexe' covers this syscall

Review URL: https://chromiumcodereview.appspot.com/9232014

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@7622 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2