Change the x86-64 decoder to store jump offsets rather than target addresses.
This change was a little complicated for two reasons.  First, it required
changing the interface for printing disassembly information so that the offset
could be transformed to an absolute address before printing.  This interface
was widely used for printing debugging information, so a large number of lines
were touched.  Second, the way that the decoder stored negative 32-bit
constants was a little off, it kept trying to treat them as 64-bit constants.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2515
TEST= ./scons platform=x86-64 validator_tests

Review URL: https://chromiumcodereview.appspot.com/9131024

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@7668 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2