Exception handling: Mac: Fix hang on unwritable exception stack
Before this change, mach_exception_handler.c would fault and hang if
it attempts to write to an unwritable exception stack.  Exceptions
inside the exception handler cause hangs, because the exception
handler is not responsive to Mach messages while it is active and
cannot be re-entered.

In the long term, we might want to make it so that page faults in the
exception handler don't hang.  In the short term, we want to avoid
this specific page fault that is triggerable by untrusted code.

We also change the Linux-ARM case to behave consistently with the
other Linux cases.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2538
TEST=exception_tests

Review URL: https://chromiumcodereview.appspot.com/9385005

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@7776 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2