Enable inbrowser_test_runner in more cases
Instead of putting the list of executables/manifest files to run in
the HTML, generate it from Scons.  This allows us to run executables
conditionally, so that a subset of the test cases can be run for PNaCl
and nacl-glibc.

Enable the test in the Chromium build.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2602
TEST=run_inbrowser_test_runner

Review URL: https://chromiumcodereview.appspot.com/9355007

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@7849 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2