NaClDesc: Rename SendMsg/RecvMsg methods to LowLevelSendMsg/LowLevelRecvMsg
This is in preparation for turning the IMC-specific functions
NaClImcSendTypedMessage() and NaClImcRecvTypedMessage() into methods
of NaClDesc so that they can be replaced by an embedder of NaCl, such
as a PPAPI proxy based on Chrome IPC.  The new methods will be called
SendMsg and RecvMsg.

The existing methods are low-level because they operate on OS-specific
NaClHandles.

The to-be-added methods will be high-level because they operate on
NaClDesc objects.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2627
TEST=compile

Review URL: https://chromiumcodereview.appspot.com/9696013

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@8015 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2