Provide a more extensible startup interface to Chromium
Rather than providing NaClMainForChromium() plus various global
setters, we use a non-global struct of arguments.  This will make it
easier to add and remove arguments in the future.

This is in preparation for adding an "enable_exception_handling"
argument.

Following the recent cleanup of SelLdrLauncher, the array of handles
is replaced with a single handle (imc_bootstrap_handle).

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2651
TEST=built Chromium locally

Review URL: https://chromiumcodereview.appspot.com/9695061

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@8019 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2