test to check pthread_cond_timedwait timeout occurs after the specified period
We received a report that pthread_cond_timedwait did not wait for the
specified duration on Linux and OSX (when not signaled), but would
instead return immediately.

This turned out to be a developer error -- the programmer was using
clock_gettime(CLOCK_MONOTONIC, &ts) and adding to the timespec to get
the absolute time for the cond timedwait timeout, which is not the
default clock used (and furthermore the condition variable was not
initialized with a condattr specifying the alternate clock -- which
would have made the error obvious since pthread_condattr_init is not
provided in the newlib-based toolchain and pthread_condattr_setclock
fails with EINVAL when invoked with CLOCK_MONOTONIC in the glibc-based
toolchain).  Even so, we didn't have an explicit test to ensure that
the timeout occurs correctly (modulo host-OS-based timing variations).
Before we had the nanosleep syscall, the library implementation of
nanosleep used pthread_cond_timedwait, so in effect we had coverage
that way, but since nanosleep was added, coverage lapsed.

This CL remedies this situation.

TEST= this is a test
BUG= None
Review URL: https://chromiumcodereview.appspot.com/10119021

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@8370 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2