Unify browser/non-browser test executables used by inbrowser_test_runner
This reduces the amount of Scons plumbing that is needed to extend a
non-browser test to run inside the browser.

Before this change, the nexes used in inbrowser_test_runner come from
nacl_env (which is non-IRT-based by default, but used for IRT-based
browser tests).  After this change, they come from nacl_irt_test_env
(which is always IRT-based).  In order to make this work, we have to
run chrome_browser_tests with "nacl_irt_test" added to --mode.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2739
TEST=run_inbrowser_test_runner --mode=dbg-host,nacl,nacl_irt_test

Review URL: https://chromiumcodereview.appspot.com/9619035

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@8439 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2