libpthread: Make pthread_t a pointer instead of a small integer ID
This removes a lot of complexity: we don't need a hash table to map
integer IDs back to pointers, so we can remove nc_hash.c.  There was
no particular need for pthread_t to be an integer ID; for comparison,
pthread_t is a pointer in glibc's libpthread.  This just changes
libpthread's behaviour when given invalid pthread_t values.  After
this change, libpthread is more likely to crash instead of returning
an error.

This is in preparation for using nc_thread.c inside the IRT so that
pthread_create() works inside the IRT.  Removing the use of integer
IDs will make it easier for IRT pthread_mutexes to work on
user-created threads since we won't have to allocate IDs for these
threads.

We add a "__" prefix to "struct nc_basic_thread_data" to avoid
namespace collisions since it now has a forward declaration in a
public header.

thread_test.c: Remove a pthread_join() check because it relies on
undefined behaviour (as previously commented in the code).

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2788
TEST=run_thread_test run_main_thread_pthread_exit_test

Review URL: https://chromiumcodereview.appspot.com/10382235

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@8650 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2