Remove most of the Windows implementation of nacl_signal
This code has never been secure on x86-64 and has never been enabled
in production.  It's a maintenance burden because it complicates
command_tester.py and it requires testing/fixing when I change the
signal handling code used for Linux.

Its only use now is printing the program counter address on a crash,
but this is better left to a debugger.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2664
TEST=run_signal_handler_test run_nacl_signal_test (on Unix)

Review URL: https://chromiumcodereview.appspot.com/10451029

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@8717 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2