Making download_toolchains more noisy by default.
We've been seeing downloads fail due to timeout on the buildbots.
It's unclear if this is due to the download taking too long, or the extract.
In fact, we're not even sure the extract step is being reached.
By making the extract more noisy, we can tell, as it has a distinct second set of dots with a separator.

Making download_toolchains verbose by default and using the verbose setting with CygTar.

BUG=None
TEST=None
R=noelallen@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10443088

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@8767 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2