An Gio subclass that captures the last N (512) bytes of output.
This Gio subclass implementation is the first step toward better error
reporting / instrumentation.

This class is intended to be used in conjunction with the NaClLog
module's NaclLogAbortBehavior to send the last few lines of log output
to the plugin (and to the JavaScript console) and to try to capture
this data in the minidump that's sent back to Google by breakpad (if
users authorized, etc).  This will hopefully both let developers know
what happened and let us know what might be going on in the field with
early, startup "command channel failed" crashes.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2776
TEST= run_nacl_error_gio_test
Review URL: https://chromiumcodereview.appspot.com/10514006

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@8832 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2