Simplify sel_main.c to remove some command-line-parsing result variables
By moving the NaClAppCtor() call earlier, we can modify the NaClApp
directly from the command line parsing loop so that we don't need
intermediate variables.

Use "nap" instead of "&state" since "nap" is the normal name for a
NaClApp* that is used in most of the codebase.

Also use LOG_FATAL for an error since it is not necessary to attempt
to clean up.

This does reduce some of the logging that will be sent to the log file
specified with "-l", but maybe the NACLLOG env var should be preferred
over "-l" since it will take effect earlier on.

BUG=none
TEST=small_tests

Review URL: https://chromiumcodereview.appspot.com/10534121

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@8917 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2