Debug stub: improve GDB tests using parsed GDB/MI output records
Instead of returning whole output from a command, pick only the record of
interest and parse it into a dictionary. Use that dictionary to perform more
accurate checking of command results.

For now, parsing is oversimiplified and error-prone, however, it works for a
number of simple cases we actually meet.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2764
TEST=run_gdb_stepi_after_break_test run_gdb_print_symbol_test

Review URL: https://chromiumcodereview.appspot.com/10699071

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9078 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2