Don't check alignment of call instructions in x86-32 validator
The checks that a call instruction appears only at the end of a
bundle are not a security/safety requirement, just a "courtesy".
For the next-generation x86 nacl-gcc, we're not generating calls
at the ends of bundles any more, so it has become discourteous to
diagnose it.

The ideal would be to look ahead and give a friendly warning if a
call appears in a bundle followed by any instructions that are
not recognized nops.  But that seems to be difficult to implement
given the structure of the validator.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2884
TEST= trybots

Review URL: https://chromiumcodereview.appspot.com/10703054

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9079 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2