Thread suspension test: Refactor how we set and check register state
Saving, restoring and comparing whole sets of registers is getting to
be a common pattern in the tests I've been writing.  It's awkward to
write assembly code to handle every register for every test, so it's
time to factor some of this out into reusable parts.

This change factors out an ASM_WITH_REGS() macro for executing a
fragment of assembly with all general purpose registers restored.

This makes RegisterSetterThread() much simpler.  Test register values
can now be set up in C.

Since we're using NaClSignalContext to represent register state, we
can pass this data to suspend_test_host.c at run time.  This means we
no longer need to duplicate the test register values across two files.

There is a risk that the testing becomes circular: if we mix up the
register ordering, we might not notice if all the code shares the same
incorrect ordering.  However, the main thing we want to check in these
tests is that we don't lose any state.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2758
TEST=run_thread_suspension_test

Review URL: https://chromiumcodereview.appspot.com/10828040

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9285 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2