Add a test for the registers reported for a thread suspended in a syscall
Add a missing "memory" constraint for the inline assembly in
register_set.h.  Without this, the compiler can optimise away writes
to the NaClSignalContext struct before ASM_WITH_REGS().

There are a couple of ways in which the registers are wrong (see the
TODOs), which I'll fix in later changes.

This change splits out AssertRegistersEqual() as a separate function
from TestGettingRegisterSnapshot().

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2758
TEST=run_thread_suspension_test

Review URL: https://chromiumcodereview.appspot.com/10830057

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9298 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2