Make the thread_create() syscall responsible for aligning the stack pointer
We want to make the system (the TCB or the IRT) responsible for
aligning the stack pointer of new threads, so that in PNaCl, user code
does not need to know about architecture-specific stack alignment
requirements.

It is easier to make the TCB responsible for this than the IRT.  The
TCB already needs to know how to align the stack when pushing an
untrusted fault handler's stack frame.

Aligning the stack in NaClCommonSysThread_Create() means we can remove
a warning from NaClCreateAdditionalThread().

We add a new test in tests/stack_alignment to check this more
rigorously than the previous tests (which don't detect alignment
problems when run using the PNaCl toolchain).

BUG=https://code.google.com/p/nativeclient/issues/detail?id=2904
TEST=run_stack_alignment_asm_test

Review URL: https://chromiumcodereview.appspot.com/10795054

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9299 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2