Give the unpacked and CWS version of the pnacl extension an "all hosts" permission to work around blockage from: https://chromiumcodereview.appspot.com/10828067/
"An extension's resources can only be loaded into a page
for which the extension has explicit host permissions".

The component updater + installer version does not have
this because we aren't using chrome-extension:// URL requests
to access it.  We then also don't need the
web_accessible_resources list for that either.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2365

Review URL: https://chromiumcodereview.appspot.com/10827143

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9361 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2