Thread suspension: Fix to report the correct return address from syscalls
Before this change, the reported prog_ctr for a thread suspended in a
syscall is the return address for the *previous* syscall (or the
thread entry point) rather than the current syscall.

Fix this by reordering the assignment of new_prog_ctr in
NaClSyscallCSegHook().  Remove some code that is duplicated across the
three architectures.

An alternative would have been to change
NaClAppThreadGetSuspendedRegisters() to read the return address off
the untrusted stack from "regs->stack_ptr - NACL_SYSCALLRET_FIX +
NACL_USERRET_FIX", but that would be inconsistent with how the other
register state is reported.  It also doesn't work for reporting the
entry point of a thread that has been suspended during startup (which
is not a case we currently test for, but I will add a test covering it
later).

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2758
TEST=run_thread_suspension_test

Review URL: https://chromiumcodereview.appspot.com/10826124

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9367 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2