Dropping debug crt copying on windows.
In the distant past, debug windows builds (built with msvs2005, ie VC80)
could be difficult to debug when moved between systems if one had an msvs
service pack installed and the other didn't. This was due to a change the in
the windows CRT. The work around, at the time, was to snapshot a copy of
the non-redistributable debug DLLs for use at debug time.

Dropping all of this as:
1. We don't build on msvs2005 anymore.
2. We don't typically debug from remote built artifacts.
(and when we do its reasonable to expect the required debug CRT will
be available or can be lived without.)

BUG=None
TEST=None
R=bsy@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10899017

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9588 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2