Use irt shims from the toolchain when enable_chrome_side=0.
We will eventually want to have the toolchain build an empty version of
libpnacl_irt_shim.a to be used by the nacl build.
The current version is built using a nacl build, from the version of
ppapi DEPSed into nacl. This suffices for now and avoid having
unusual parameters passed to linker / translator.
(This is a slight circularity, but is temporary.)

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2957
TEST=None
R=jvoung@chromium.org,ncbray@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10894044

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9620 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2