Create a "sel_main_chrome" Gyp library for Chromium to depend on
This will allow the Chromium Gyp files to be cleaned up so that they
do not depend on "sel" or "platform_qualify" directly but instead just
depend on this single entry point target.

Also change the Scons build similarly and change it to reflect the
real dependencies between components.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2832
TEST=build

Review URL: https://chromiumcodereview.appspot.com/10909069

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9633 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2