Add nacl_sdk gyp meta-target.
This is a part of the making the nacl_sdk builder use gyp instead of scons. The
libraries built should be the same as the libraries build when building the
scons install_lib target, excepting the ppapi libraries which will be built on
the Chrome side.

BUG=none


Review URL: https://chromiumcodereview.appspot.com/10915109

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9663 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2