CreateProcess SRPC call added to Reverse Service.
This call will be used to spawn new sel_ldr processes. The call
has no input argument and the only output argument is the
socket address descriptor of the secure command channel which
can be used to set up the newly spawned process.

This call is not exposed through the wrapper to the plugin at
the moment as the initial implementation is assumed to be
for standalone sel_ldr.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=3019
TESTS=small_tests

Review URL: https://chromiumcodereview.appspot.com/10915193

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9774 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2