x86-64: Remove the nacl_tls array and use NaClThreadContext instead
Now that we have symbolic constants for offsets into
NaClThreadContext, it's easy to do this.

The logic we use in NaClThreadContextCtor() to avoid clobbering tls1
mirrors the ARM case.  We add a just-in-case memset() call to mirror
the ARM case as well.

Removing nacl_tls[] saves 32k of memory on all architectures.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2881
TEST=run_hello_world_test run_thread_test

Review URL: https://chromiumcodereview.appspot.com/10831083

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9829 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2