Tweak x86-32 testdata file for unaligned_data_irt test
The testdata binary needed for https://codereview.chromium.org/10983053
conflicts with an outstanding regression in the Ragel validator vs the
production validator.  The difference is irrelevant for this test, so
just use a different binary rather than waiting for the Ragel validator
to be fixed.

TBR=bsy@google.com

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2814
TEST= none

Review URL: https://codereview.chromium.org/10985066

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9857 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2