Debug stub: Ensure that tests run on the bots for ARM
The debug stub tests are no longer getting run on the PNaCl bots for
ARM.  This was probably a result of r9810, which changed how
large_tests got run.

Before this current change, large_tests couldn't safely be run with
-jN (N>1) because it contains GDB tests which bind a fixed TCP port
and so can't be run concurrently.

We address this by using Scons' SideEffect() call to declare that the
tests should be run sequentially, so using -j on the tests will now be
safe.

(Note that SideEffect() is already used this way by
src/trusted/validator_x86/build.scons.)

Add run_debug_stub_test to nonpexe_tests so that it gets run on the
ARM bots.  Move run_debug_stub_test to medium_tests because, with the
concurrency issue resolved, it is not sufficiently resource-intensive
to be considered "large", and because large_tests might not be run
fully for ARM.

Move "tests/debug_stub" to nonvariant_tests, so that we don't attempt
to run it with the IRT.  (It was not run with the IRT before, because
large_tests_irt isn't run on the bots.  It also doesn't pass with the
IRT.)

Add the "gdb_tests" alias for convenience.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2911
TEST="large_tests -j4" passes without failing to bind an occupied port

Review URL: https://codereview.chromium.org/11086031

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@9974 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2