service_runtime: Move TLS values to live exclusively in NaClThreadContext
This removes an arch-specific #if for x86-32 from nacl_app_thread.h.
The layout of the %gs segment is now specified in the arch/x86_32
directory.

On x86-64, this removes duplication, so that tls_value1 is stored in
only one place.  This will also allow adding a fast path syscall for
tls_value2.

On ARM, this will allow us to change nacl_syscall.S to use r9 rather
than a __thread variable to retrieve the current NaClThreadContext.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2881
TEST=run_thread_test, run_irt_thread_test, etc.

Review URL: https://codereview.chromium.org/11094038

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10003 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2