For the archived-frontend-test, build the IRT with the latest toolchain.
This is more like reality (in that the IRT constantly changes) before
pulling down the archived frontends to build the test code.

More specifically, the ARM IRT is now built with a biased bitcode triple,
which the old frontends do not know how to produce.

Also use separate BUILD_STEPs to make it more clear if things break
during the build or during the run step.

BUG= red bots

Review URL: https://codereview.chromium.org/11175008

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10050 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2