Gyp build: Enable -Wundef warnings for untrusted code
-Wundef warns about "#if FOO" when FOO is undefined.

We have to ensure NACL_LINUX, NACL_WINDOWS etc. are defined, because
various header files check them.

We don't want to enable this for Chrome source because that produces
errors such as:
  ../base/logging.h:187:5: error: "NDEBUG" is not defined
  ../base/pending_task.h:21:5: error: "_MSC_VER" is not defined

Also enable -Werror for ARM untrusted code built with Gyp inside the
NaCl tree so that we catch -Wundef problems in this case.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2787
TEST=Gyp build + tested in Chromium

Review URL: https://codereview.chromium.org/11262013

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10089 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2