Fix alignment constants for pnacl assembly code.
Assembly code processed by Pnacl's assembler should not rely on magic
symbols injected by the assembler for alignment, but define them
explicitly. This is already done in the ARM version of setjmp. This CL
modifies the x86 versions to follow suit.

Also see Issue #3115 for more details. This CL implements step 1.

Note: currently there's some amount of duplication between various
instances of nacl_config.h headers defining NACL_BLOCK_SHIFT constants.
It exists in the ARM version, and this CL extends it to the x86 version.
It would make sense to clean this up globally, but should probably go
into a separate CL in order to not obscure the goal of this one.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=3115

Review URL: https://codereview.chromium.org/11336025

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10133 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2