Make certain "meta" command line args in the pnacl drivers work when a driver invokes another driver.
This fixes:
-pnacl-driver-set-XXXX and -pnacl-driver-append-XXXX

Currently those flags are broken for the recursive invocation case
because they are passed through and can cause asserts
when the XXXX variable is not defined in the recursively called driver.

Also, moved a flag in the translator into a variable
and fixed the the no-shim option.


BUG=


Review URL: https://chromiumcodereview.appspot.com/11362204

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10244 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2