Performance tests: Print more digits for timing results
Print sub-nanosecond times because some tests take on the order of
nanoseconds per iteration, and so we have been printing only one
significant figure for the per-iteration times.  This meant that the
variation in the results is not revealed in the graphs on Buildbot,
which either show a single line for these tests, or a line that
oscillates between two values.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3161
TEST=run_performance_test

Review URL: https://codereview.chromium.org/11308312

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10361 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2