Performance tests: Fix TestThreadWakeup, which was accidentally spinning
The logic in TestThreadWakeup's child thread was wrong: After setting
"state = REPLY_TO_PARENT", the wait loop's condition "state == WAIT"
is false so we immediately fall through to setting "state =
REPLY_TO_PARENT" again.  In the worst case, the parent rarely gets a
chance to run because the child thread is claiming the mutex so
frequently.

I realised this test was wrong after trying to switch nacl-newlib to
using futex-based mutexes and condvars and getting a huge slowdown in
this test.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3161
BUG=https://code.google.com/p/nativeclient/issues/detail?id=562
TEST=run_performance_test

Review URL: https://codereview.chromium.org/11421217

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10374 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2