Change gc_instrumentation test to test IRT interfaces instead of libpthread
The behaviour of both the IRT and libpthread is subject to change
regarding when they generate blocking-notification callbacks.  But the
IRT is the component that is reponsible for generating the callbacks,
so we test the IRT rather than libpthread, because libpthread is one
level removed.

This is in preparation for changing nacl-newlib's pthread_mutex() to
use futexes, which means it won't generate blocking-notification
callbacks in the uncontended case.

This test was not run under nacl-glibc before, but after this change
it no longer links with nacl-glibc (because of the use of
__nc_irt_mutex so that this continues to work in the non-IRT case).
This means we have to move the test-disabling logic.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=562
TEST=run_gc_instrumentation_test

Review URL: https://codereview.chromium.org/11440002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10385 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2