Thread suspension: Fix querying registers for another Mac OS X corner case
When we resume an untrusted thread with registers modified on Mac OS
X, NaCl makes the thread run a small trusted-code routine to restore
x86-32 segment registers.

There is a small chance that the thread is subsequently suspended
while it is still executing that routine.  If that happens,
NaClAppThreadGetSuspendedRegisters() needs to reconstruct the register
state appropriately.

The routine consists of two parts:

 1) NaClSwitchRemainingRegsViaECX
 2) NaCl_springboard_all_regs

NaClAppThreadGetSuspendedRegistersInternal() already handled (1); this
change fixes it to handle (2) as well.

This is difficult to test.  I haven't been able to reproduce the
problem locally.  I tried writing a stress test, but it didn't
reproduce the problem.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=3240
TEST=no new test, just gdb_tests and run_thread_suspension_test

Review URL: https://codereview.chromium.org/11759016

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10535 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2