Adjust pnacl-translate sbtc flag override mechanism to append.
Separate -mcpu from the default, so that it can be
overridden. Remove LLC_FLAGS_SLOW.  That is considered the
default right now, but it was never put into the default
commandline flags in the sandboxed translator, so it's not
really default anyway.

Also test that the overrides actually happen (especially
for the RPCs).

Depends on: https://codereview.chromium.org/12459004/

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3325

Review URL: https://codereview.chromium.org/12459004

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10971 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2