build_nexe.py: force rebuilds when the toolchain changes
The most recent PNaCl toolchain has an incompatibility between old bitcode files and new translators, which is causing the ARM IRT build to fail because build_nexe.py is not forcing a rebuild of the object files on the bot. This CL adds an extra check of the toolchain prep stamp and forces rebuilds when the toolchain is newer than the output.

R=sbc@chromium.org
BUG=none (blocking roll of NaCl into Chrome)

Review URL: https://codereview.chromium.org/12907024

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@11076 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2