Use gcc by default for gyp untrusted ARM builds.
This is needed to get working .nexe test files out
of a chromium ARM build.

The old (broken) path can still be enabled using the
new force_arm_pnacl option.  This path is needed to
build the pnacl ARM shims.

This required a chrome-side change to land there
and not break the build:

https://codereview.chromium.org/13129002

BUG=196627
TEST=run nacl brower tests one ARM linux

Review URL: https://chromiumcodereview.appspot.com/12499011

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@11120 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2