Implement torture test script in python using buildbot_lib
The current script, torture_test.sh doesn't properly check the return value
on its calls to eh_tests and standard_tests.
This script fixes that, as well as using the buildbot_lib and standard python
stuff to make the script a lot shorter and no more complex,
and is structured very closely to torture_test.sh

I'd like to do this in 2 stages: first this rough draft, just to get the
bot reporting the results correctly; and then cleanup and perhaps better
integration with buildbot_lib, and eventual replacement of
buildbot_pnacl_toolchain_tests.sh with a python script as well.
So for this round, please just major stuff
(although feel free to leave comments here for the next round)

BUG=none
TEST=torture test bot

Review URL: https://codereview.chromium.org/9737002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@11194 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2