Update third_party DEPS to pull bloat and webtreemap.
They were added in:
  https://codereview.chromium.org/14746016/

These will be used in our infrastructure to track the filesize of the llc.nexe that's shipped with Chrome (it's currently way too big at ~11MB). The goal is to track it, know when we're wasting space, and identify regressions (especially when we do upstream merges).

It will only be used by the bots, and the html+css+json files will be archived.

R=dschuff@chromium.org
BUG= none
TESTS= none

Review URL: https://codereview.chromium.org/14672015

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@11324 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2