build_nexe.py now normalizes paths before accessing the filesystem.
Without this change, on Windows, a long relative path will test as
non-existent. Even if the relative path is < 255 characters, and the absolute
path is < 255 characters, if the joined CWD and relative path are > 255
characters, the file cannot be opened or stat'ed.

The workaround here is to lexically normalize the path before any filesystem
access.

BUG=none
R=sbc@chromium.org

Review URL: https://codereview.chromium.org/15732011

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@11403 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2