[MIPS] Remove NaClCheckDEP for MIPS
MIPS sandboxing does not require DEP, and the target MIPS platforms do not
support it. Thus, this change removes the code we do not need.

NOTRY=true
However, we have run a try job manually, and it passed fine, except for
the nacl-lucid64_validator_opt that is broken.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=2275
TEST= run smoke tests for MIPS platform

Review URL: https://chromiumcodereview.appspot.com/18861003

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@11744 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2