Test that calls farther than 32MB away work, update ARM toolchain
On ARM, calls to targets more than 32MB away require special linker
support.  The linker generates code that's effectively a load and an
indirect branch, so that code needs to meet NaCl sandbox requirements.
The test applies to non-ARM platforms as well because far calls need
to work everywhere.  It just so happens that x86 doesn't have the
branch-distance limitations that necessitate the complexity in which
the ARM bug existed.

The ARM toolchain update brings in a linker that has this issue fixed.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3559
TEST= trybots
R=mseaborn@chromium.org

Review URL: https://codereview.chromium.org/18804007

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@11818 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2