Update NaCl perf expecations for nacl-gcc (from change at 11467).
Commit r11467 (rudimentary asm.js support) refactored the setup of
the input files into a python script.  This setup is actually
measured as part of the timing.  It appears that in some cases
this leads to a noticeable runtime increase and pushes the runtimes
to the boundaries of the perf expectations.

BUG=none
R=stichnot@chromium.org

Review URL: https://codereview.chromium.org/20049002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@11835 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2