Add --allow-llvm-bitcode-input to pnacl-translate.
Without it, pnacl-translate will refuse to translate LLVM IR, thus being more in line with the sandboxed translator. This will hopefully provide a better user experience for developers.

Other fixes are in other drivers and tests that invoke pnacl-translate -- they should now pass the new flag whenever dealing with LLVM IR.

A precursor commit was pushed (r11938 in NaCl SVN) adding the option as "recognize but ignore" to pnacl-translate, so that the toolchain bots build a toolchain with this in. This way future try jobs and bot runs for this CL should not fail.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3519
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/22470002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@11944 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2