Thread suspension: Extend regs_step_test to work on ARM
regs_step_test single-steps through a NaCl syscall invocation to check
that thread suspension reports the correct register state at each
instruction.

This is harder to do on ARM than on x86, because ARM lacks hardware
support for single-stepping.

It is possible to emulate single-stepping on ARM by setting a
breakpoint on the next instruction that we think will be executed.
That's what GDB does.  However, this requires having a partial ARM
instruction decoder.  This is fairly easy for original ARM, but harder
for Thumb-2 (especially If-Then instructions), and NaCl trusted code
includes Thumb-2 code (e.g. in libc).

So we take a simpler approach and set breakpoints on whole ranges of
instructions that we're interested in.  We remove and reinstate these
breakpoints one at a time in order to step through this code.

Add NaClSyscallSegEnd so that we can cover the NaClSyscallSeg routine
with breakpoints.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=2830
TEST=run_signal_handler_regs_single_step_test

Review URL: https://codereview.chromium.org/22290011

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@11956 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2