Run -expand-byval to make calling convs match, when not using --target=*
The current way of avoiding ABI simplification passes
completely when using C++ exceptions works for NaCl only
programs, but it does not work for PPAPI programs.
That is because PPAPI programs link to a shim which expects
us to have changed the calling convention.

Run -expand-byval to make calling convs match for PPAPI
programs.  Don't run it if targeting the native ABI.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=3488
NOTRY=true
R=dschuff@chromium.org

Review URL: https://codereview.chromium.org/22867042

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@12072 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2