Add OWNERS files for src/untrusted/{irt,nacl,pthread}
Changes to IRT interfaces should require review from owners because
these are the long-term stable interfaces that NaCl is committing to.

The libnacl and libpthread code can also be fairly subtle.  We don't
want this code to make incorrect assumptions about the interfaces the
IRT provides, or do things that are not strictly thread safe.  So add
OWNERS files for these too.

BUG=none
TEST=none

Review URL: https://codereview.chromium.org/58633003

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@12359 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2