toolchain_build: Build GDB
Bring the build of nacl-gdb into the toolchain_build universe.  This build
is not yet used by anything, and the old build remains in tools/Makefile.
This build uses a locally-built expat that is linked statically into GDB.

Note that this new build uses nacl-gdb/* branches in the nacl-binutils.git
repository, abandoning the old nacl-gdb.git repository.  The branch
nacl-gdb/7.7/master in the new repository is a close match for the tip of
the old repository's master branch, but organized so that merging in
upstream changes will be simple.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3739
TEST= toolchain trybots
R=bradnelson@google.com

Review URL: https://codereview.chromium.org/148173013

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@12675 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2