Adds tls_edit utility which patches irt_core.nexe's TLS usage.
Currently we compile in different instructions to manage the 2 TLS
usages between an untrusted nacl app and the IRT. This adds a new
utility (tls_edit) which goes through nexe's and edits the
instructions which use the TLS to point to the correct location.
This way we don't have to do anything special when compiling the
IRT.

The ComponentProgram function has also been separated so creating
of the ComponentProgram is done separately from aliasing programs
to the "COMPONENT_PROGRAM_GROUPS" aliases. This way we can add in
extra steps between the building of the Component Program and
adding the program to the correct component program aliases. This
is now necessary because the IRT core is now built and uses
tls_edit as an intermediate step before the nexe is complete.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3127
TEST= trybots
R=bradnelson@google.com, mcgrathr@chromium.org, mseaborn@chromium.org

Review URL: https://codereview.chromium.org/140653005

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@12698 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2