Remove the unused and non-functioning ioctl syscall
This syscall is not hooked up to do anything -- there is no
implementation of NaClDesc's Ioctl() method that does anything.

It is never called -- there is not even a syscall wrapper for it -- so
it's better to remove it.

If we want to expose any interfaces that are normally exposed via
ioctl() on Linux, such as isatty(), it will be better to provide these
via specific syscalls (and functions in IRT interfaces) rather than
via ioctl().  As a catch-all, ioctl() is awkward to vet for security
because there is no uniformity in how individual ioctls access the
calling process's memory.

BUG=none
TEST=trybots
R=bradnelson@google.com, sbc@chromium.org

Review URL: https://codereview.chromium.org/153473002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@12850 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2