Uploads a package using package_version for all the toolchain builds
As we are transitioning to the new package versioning tool, we need
to first have a package available for all of the different toolchains
for each SVN revision. Once we have a couple of packages uploaded, we
can set the sanctioned revision using package_version.py.

BUG=
TEST= trybots
R=bradnelson@google.com, jvoung@chromium.org, noelallen@chromium.org

Review URL: https://codereview.chromium.org/200543008

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@12982 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2