ARM: Avoid leaking an address of trusted code to untrusted code
Before this change, the address of NaClSyscallSeg (trusted code) is
visible in the trampoline code.  This defeats address randomisation of
trusted code.

We address this by putting the address of NaClSyscallSeg into
NaClThreadContext and by fetching this address via r9.

The simplifies the trampoline setup code, because the trampoline code
no longer requires any patching: each trampoline is the same.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=3043
TEST=run_hello_world_test run_nacl_thread_capture_test
R=mcgrathr@chromium.org

Review URL: https://codereview.chromium.org/224903003

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@12997 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2