Change PNaCl --noirt use of rosegment-gap to something bigger.
rosegment-gap measures the gap from the start of text not
the end of text. E.g., the default for NaCl is 0x10000000.

So, for it to take effect on --noirt builds it needs to be
a bit more generous that 32 bytes. This isn't a problem
for now, because PNaCl's linker doesn't pad out the text
segment to the end of a page, but will once we merge
binutils up any further than it is now. Once that happens
there won't be room for the halt sled.

Set to 12MB for now which should be more than sufficient
for pnacl-llc.nexe and ld.nexe.

BUG=none
R=dschuff@chromium.org

Review URL: https://codereview.chromium.org/247783002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13083 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2