Add chromium_git variable to fix sync
I have added the first git repo pinned for native client.  In the chromium
DEPS they defined the "chromium_git" variable, but this was not previously set
here.  Set it, in the hopes that this will put out the forest fire and make
the .DEPS.git change work this time.

BUG=
R=bradnelson@google.com

Review URL: https://codereview.chromium.org/289003002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13184 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2