Make android build with -Werror
There were a number of issues with -Wformat that needed to be fixed.  Notably Android's definitions of size_t and ssize_t were different than any platform previously supported by nacl (one is an int and one is a long).  And there were several uses of PRIuS with ssize_t arguments and PRIdS with size_t, etc.
One other small issue was with strerror_r's return value, which is different as a gnu extension (it returns the buffer) from elsewhere (it returns an integer status code).  Since the client of that function was not checking the error return, I simply removed reliance on the return value.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3032
TEST= hello_world_arm_newlib.nexe under adb shell
R=mseaborn@chromium.org

Review URL: https://codereview.chromium.org/292153005

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13223 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2