Don't remove -m32 when building ia32 targets.
There used to be a host_arch conditional that got lost when munge_phdr moved
from chromium to nacl, see
https://codereview.chromium.org/8800034/diff/4001/chrome/nacl.gypi
https://codereview.chromium.org/8623002/diff/10001/src/trusted/service_runtime/linux/nacl_bootstrap.gyp

Keeping the -m32 flag when target_arch and host_arch are both ia32 is important
when building with a 32bit userspace on a 64bit kernel using a 64bit clang
binary as compiler.

As far as I can tell, -m32 and -sysroot are only added for target binaries,
so this hack shouldn't be necesesary anymore. (And if it turns out that it
still is for some reason, that reason can likely be fixed instead.)

(Somewhat related to https://codereview.chromium.org/342493002/)

BUG=368384
R=bradnelson@google.com

Review URL: https://codereview.chromium.org/341573003

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13369 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2