pnacl-ld: Do not attempt to parse ELF or bitcode files as linker scripts
Currently ldtools attempts to parse a file to see if it is a linker
script. An ELF file was found that confused the parser and caused it to
try to seek outside the file. Do not try to parse files that are known
to be non-scripts.

Also move ParseLinkerScript and IsLinkerScript to filetype.py
to break the circular dependency.

R=jvoung@chromium.org
BUG=none

Review URL: https://codereview.chromium.org/379353002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13466 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2