Disable flaky sandboxed translator tests on mac
The PNaCl offline sandboxed translator is flaky on mac and causes the
tree to close unnecessarily. Disable its tests for now.

Also add the mac trybot to the default trybot set now that it
won't be flaky.

Also, since the sandboxed translator tests are really just to test that
the translator runs, and not to cover all the possibilities for input,
it's not really necessary to run the full nacl test suite.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3856
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/401073003

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13510 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2