[MIPS] Disable the PNaCl ABI verifier when building the IRT
__executable_start is not part of PNaCl's stable ABI, and
PNaClABIVerifyModule will report that it is not a valid external symbol.
So, disable this check when building the IRT. This fixes a regression on
MIPS caused by r13514.

NOTRY=true
However, I have run a try job manually, and it passed fine, but a few
bots are flaky, so this is a workaround for it.

BUG= irt_core can not be built for MIPS
TEST= ./scons platform=mips32 irt_core

Review URL: https://codereview.chromium.org/423313002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13563 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2