Added new 'pnacl-llc' driver
This adds a new driver into the SDK's bin directory, which invokes the native
LLVM static compiler needed to compile MinSFI binaries. The existing 
'pnacl-translate' driver could be used but it is mainly intended to target NaCl, 
which complicates native compilation of code that is not PNaCl ABI-compliant.

NOTRY=true
BUG= https://code.google.com/p/chromium/issues/detail?id=397177
TEST= manual check that it is installed and that it passes args correctly

Review URL: https://codereview.chromium.org/540503005

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13694 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2