Add a get_random_bytes() syscall to replace the SRPC-based implementation
This will let us remove one remaining use of SRPC.

This allows us to simplify tests/random/, since it no longer needs to
link in a copy of libsrpc.

Note that this doesn't remove "SecureRandom" from the name service or
remove nacl_desc_rng.c.  I'll do that in a follow-on change in order
to keep this change smaller and easier to review.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3864
TEST= run_random_test run_random_test_irt

Review URL: https://codereview.chromium.org/537543003

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13696 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2