Build x86-32 IRT with PNaCl
This builds the IRT with PNaCl instead of nacl-gcc on x86-32.
This uses the same flags and build style that we use on x86-64,
and will allow use of C++11 language features in the Chromium
code that goes into the IRT. Some notes:
 * the x86-64 IRT already uses PNaCl.
 * the ARM IRT uses arm-nacl-gcc which is gcc 4.8 and so
 has C++11 support. We plan to switch ARM to PNaCl as
 well anyway for consistency, and to simplify the GYP.
 * PNaCl still uses libstdc++ instead of libc++ for the
 IRT. This was done for consistency with the gcc builds,
 but we will switch to libc++ soon. This will allow C++11
 library features as well (and will allow us to drop
 libstdc++ from PNaCl entirely).

This is a reland of r13687/13692, with the exception that we still
build with nacl-gcc in scons on Windows, and disable SSE instructions.

R=mseaborn@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3910
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3935
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3936

Review URL: https://codereview.chromium.org/557233004

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13733 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2