Cleanup: Use NaClCopyInFromUser() in one case in mmap()
Using NaClCopyInFromUser() is preferred over NaClUserToSysAddrRange()
(+ a volatile memory access) where possible, because it takes care of
any necessary locking.

It looks like this case was missed when the NaClCopyIn/Out functions
were introduced.

Also remove an unnecessary check for "offp == 0".  This is left over
from when mmap() was converted from taking a 32-bit offset arg (by
value) to taking a 64-bit offset arg (by pointer).  It was added as a
sanity check against accidentally forgetting to convert call sites of
the mmap syscall.  The conversion was done long ago, so this check is
no longer needed.

BUG=none
TEST=run_mmap_test and others

Review URL: https://codereview.chromium.org/570653002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13758 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2