allow avx/xop/fma instructions but disable all them via def file attributes.
introduce a def file attribute (disable_untested) which means that an
instruction is disabled because it is untested. As we test these instructions
we will remove this attribute, which will enable them. Changing the DFA to explicitly disallow vmaskmovdqu (which was previously the only AVX instruction in the DFA but disabled via cpufeature_avx which is now allowed).

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3327
R=khim@google.com, kschimpf@google.com

Review URL: https://codereview.chromium.org/662463002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13947 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2