Build gold.nexe with a segment gap, and supply -B, to later support IRT use.
This does not yet *require* the IRT, since the private libs
are still linked in first (need a change to the gold
Makefile). Also update the driver to start supplying the -B
flag for the IRT for sel_universal when running gold.nexe.

After that will be the actual change to stop using
the private libs, but we need to make sure Chrome is
ready and supplies the IRT first.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3914
R=dschuff@chromium.org, mseaborn@chromium.org

Review URL: https://codereview.chromium.org/668723002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13950 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2