Allow vzeroall and vzeroupper in ragel dfa.
These are AVX instructions that have no operands. They respectively zero out all the YMM registers or just the top 128 bits of all the YMM registers. The corresponding test
for adding these are in vzero_proof.py, where we verify that no instructions were removed from the DFA, and that the set of added instructions to the DFA disassembles to {'vzeroall', 'vzeroupper'}

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3327
R=kschimpf@google.com

Review URL: https://codereview.chromium.org/660473005

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13964 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2