Add a whole lot of new stub functions to libnacl.
These are the functions that we currently stub in
the naclports glibc-compat library.  Pushing these
upstream into libnacl means more software will
compile out-of-the-box and allow us to remove this
extra layer from naclports.

Many of these functions are already declared in
the newlib headers but lack implementations. Others
are not declared in the headers yet but we plan
to add them (currently extra system headers are
added by the SDK and also by naclports).

A few of these are added outside the 'stubs' folder
as we can provide meaningful implementations (not just
ENOSYS):
 eaccess
 htonl/ntohl/ntohs/htons
 sigblock/sigsetmask/sigmask/sigprocmask/sigsetmask

This change also added unittests for these non-stub
functions.

Removing stub for execlp().  This function is already
provided by newlib (libcrt_common.a) and ends up
calling _execve (which we already provide a stub for).

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3973
TEST=run_syscall_test run_inet_test
TEST=make all in naclports with stubs.c removed.
R=binji@chromium.org, mseaborn@chromium.org, noelallen@chromium.org

Review URL: https://codereview.chromium.org/642873002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@14027 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2