Add neg nan to math intrinsics/libm func test, and print nan vs -nan.
Add -NAN to the set of test values. This is in preparation
for testing fabs from libm vs the fabs intrinsic, and seeing
the behavior on -NAN.

Also, it turns out that newlib doesn't print "-nan" when
given a negative nan, and just prints plain "nan". See the
following which makes no attempt to test the sign bit:
https://chromium.googlesource.com/native_client/nacl-newlib/+/master/newlib/libc/stdio/vfprintf.c#1229
So, we wouldn't have been able to tell the difference.

Make a custom print routine, which is somewhat useful for
papering over some differences libc and differences noticed
by this extra test case (sqrt, sin, cos).

BUG= https://code.google.com/p/nativeclient/issues/detail?id=4031
BUG= https://code.google.com/p/nativeclient/issues/detail?id=4038
BUG= https://code.google.com/p/nativeclient/issues/detail?id=4040
R=jfb@chromium.org

Review URL: https://codereview.chromium.org/793463003