Filter non-digit characters out of archive member name offsets in artools
artools.py parses archive headers to determine the archive type. There are
a few special cases for long member names which encode the offset into
a table rather than the name itself in the member name field. In that case,
filter non-digit characters out of the field when converting to an int.
This matches the behavior of gnu ar and handles some cases that caused
failures previously.

R=jvoung@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=4053

Review URL: https://codereview.chromium.org/853863004