compiler_rt build: pass in ARCH
I'm writing a CL for compiler_rt which specializes which of the runtime
functions gets included into each of the architectural targets. The easiest way
to do this is to pass in an ARCH variable, instead of doing e.g. what
clang_linux.mk does by shelling out to CC and grepping for 'Target:'. This CL
can go in before the compiler_rt change because it should be a noop, until the
compiler_rt change requires it.

A future CL could further specialize by treating the nacl-clang build of
compiler_rt separately from the pnacl-clang build.

R=dschuff@chromium.org, jvoung@chromium.org
BUG= none

Review URL: https://codereview.chromium.org/997813005