Non-SFI mode: Implement {get,set}rlimit() syscall wrappers.
This is for the preparation of Non-SFI sandbox implementation.
getrlimit() and setrlimit() are used in both base/ and sandbox/.
So, having these functions in PNaCl toolchain would be beneficial.

Note that RLIMIT_ prefixed flags are limited to ones needed to
build nacl_helper_nonsfi.

TEST=./scons run_rlimit_test bitcode=1 nonsfi_nacl=1 with and without platform=arm. Ran trybot.
BUG=http://code.google.com/p/chromium/issues/detail?id=358465

Review URL: https://codereview.chromium.org/998233002