Update and triage gcc torture known failures
Fix paste-o in clang known failures to get the correct set. Also some triage
of the new failures with comments in the exclusion file, and I also found that
previous efforts in the assembler caused one test to be fixed for PNaCl as
well as clang.

R=stichnot@chromium.org
BUG=none

Review URL: https://codereview.chromium.org/1065663002