Make thread_test other pthread_mutex_unlock test POSIX-compliant
The test for calling pthread_mutex_unlock on a mutex that was
never locked invokes undefined behavior, which is why glibc
doesn't diagnose the error.  This test is only kosher with a
PTHREAD_MUTEX_ERRORCHECK mutex, which POSIX requires to diagnose
this error gracefully.

BUG= none
TEST= trybots
R=phosek@chromium.org, mseaborn@chromium.org

Review URL: https://codereview.chromium.org/1119973002