Revamp list_mappings test of phdr mappings
The test was written in C but compiled as C++.  Build it as C
instead.  The only code changes for that were some cosmetic trivia.
Because of the file rename, diffs of this change will be misleading.
The test_list_mappings_phdrs is basically completely rewritten, and
the rest of the file is not changed significantly.

The test_list_mappings_phdrs code is completely revamped.  The old
code tested that each phdr falls completely inside one mapping.
This was not the right semantics to verify, as it's entirely normal
for one phdr to wind up split among multiple mappings of contiguous
virtual pages.  The old code's test for page protections matching
was quite fuzzy.

The new code tests that every PT_LOAD phdr is covered by one or more
mappings.  The page protections matching is made precise, with extra
logic to handle the PT_GNU_RELRO case wherein some pages will have
been mprotect'd to read-only in a PT_LOAD segment that is read-write.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3068
TEST= trybots
R=bradnelson@google.com, phosek@chromium.org

Review URL: https://codereview.chromium.org/1129623002