NaClDescQuota: Remove unused serialisation methods
The code for serialising NaClDescQuota over IMC is never used because
this descriptor type (which is used by the Chromium side) is never
sent across IMC/SRPC.  We can therefore remove it.

This makes it clearer that the NaClDescQuotaInterface argument to
RecvMsg() and NaClDescInternalize() is unused and can be removed.

BUG=none
TEST=ran a Chromium trybot

Review URL: https://codereview.chromium.org/1149183003