Remove unused untrusted implementations of NaClSendDatagram() and others
In trusted code, NaClSendDatagram() and NaClReceiveDatagram() are part
of the internal layers of IMC, on top of which higher-level NaClDesc
passing is implemented.  It does not make sense to have untrusted
implementations of these, and they are not used anyway.  Their
presence was confusing to someone new to this code (see
https://groups.google.com/d/msg/native-client-discuss/5B5LFptZTyM/fI48VK4PDhUJ).

We can also remove some other unused definitions from this file.

BUG=none
TEST=build

Review URL: https://codereview.chromium.org/1210053007