Add --no-nacl-gcc option toolchain_build_pnacl.py
Currently we require the nacl-gcc toolchain in order to build
a working libgcc_eh on x86.  This options forces the use of
nacl-clang instead which results in a broken libgcc_eh but
allows the PNaCl toolchain to be built without the nacl-gcc
toolchain, which is very useful for people that want to
bootstrap from source.

BUG=https://code.google.com/p/chromium/issues/detail?id=510985
BUG=https://code.google.com/p/nativeclient/issues/detail?id=1933
BUG=https://code.google.com/p/nativeclient/issues/detail?id=3954

Review URL: https://codereview.chromium.org/1237263006