buildbot: Avoid goma for Gyp compile inside toolchain build
When a toolchain builder runs the Gyp compile to test its freshly-built
NaCl toolchain, that toolchain is never going to be available on the
goma servers.  The fallback to running locally means the bot winds up
trying to do an all-local build with ninja -j50, which can overload the
machine.

BUG= none
TEST= none
R=ncbray@chromium.org, bradnelson@google.com

Review URL: https://codereview.chromium.org/1257773005 .