Make irt_pnacl_translator_compile work without needing SRPC or sel_universal
This implementation of the IRT interface is used mainly for the
purpose of testing PNaCl's sandboxed translator in the standalone NaCl
build.  This change makes it work without sel_universal, using just
sel_ldr.  Args are passed via env vars.

As before, we can't change the IRT and pnacl-translate.py in
lock-step: the non-toolchain-rebuilding buildbots will only get the
IRT change.  So we have to keep the SRPC code until the PNaCl
toolchain revision is rolled.

 * Move some code in order to share it between files:
    * Factor out AddListToEnv().
    * Factor out SelLdrCommand().
    * Move env_var_prefix_match() into a separate file.

 * sel_universal is now no longer used by the PNaCl tools, so remove
   references to it.

 * Remove the "-bitcode-stream-rate" option for now, since we don't
   test it anywhere.

 * Update the driver scripts' tests: Now check for env vars instead of
   "StreamInitWithSplit".  Use "re.S" for the regexp match because
   some of the output from "--pnacl-driver-verbose" is now split
   across lines (with "\").

BUG=https://code.google.com/p/chromium/issues/detail?id=302078
TEST=run_hello_world_test use_sandboxed_translator=1

Review URL: https://codereview.chromium.org/1266693004