Remove src/untrusted/crash_dump/, which was superseded by minidump_generator
The crash_dump library was an early stop-gap solution for generating
crash dumps.  It has been superseded by generating Minidump-format
crash dumps and using Breakpad to decode them.  We can remove it to
reduce the amount of code we need to maintain, and to remove any
confusion about which library we recommend.

BUG=none
TEST=trybots

Review URL: https://codereview.chromium.org/1288333003