Revert of win: Update to Windows 10 SDK (patchset #1 id:1 of https://codereview.chromium.org/1389763002/ )
Reason for revert:
Chrome still has some Vista builders that can't run the Win 10 SDK tools.

So I'm going to revert this, and then reland it when I try to reland on the
Chrome side so that NaCl remains rollable in the interim.

Original issue's description:
> win: Update to Windows 10 SDK
>
> Chrome-side update here: https://codereview.chromium.org/1378153003/
>
> Make defines consistent in gyp build. GN will be updated when a
> new //build is rolled for nacl standalone.
>
> R=ncbray@chromium.org
> BUG=491424
>
> Committed: https://chromium.googlesource.com/native_client/src/native_client/+/9c69f1fcb35b34f136da7ea7c91ceb743b9d2aaa

TBR=ncbray@chromium.org
BUG=491424

Review URL: https://codereview.chromium.org/1391783002