GN: Use special toolchain for NaCl assembly code on Windows
The trusted assembly code (all .S files) in NaCl needs to be
built with the GNU assembler on Windows, rather than with the
MSVC assembler.  Doing this under GN requires using a special
toolchain, which in turn means that each target containing some
assembly files must be split out into a source_set specifically
for the assembly code so it can be referred to with the qualifier
for the special "nacl_win_as" toolchain.  Using source_set this
way in turn requires that the static_library targets be converted
into source_set targets, because static_library does not play
well with source_set as a dependency.

To simplify the changes, some small BUILD.gn files in subdirectories
are folded into the src/trusted/service_runtime/BUILD.gn file.

BUG= 512869
TEST= trybots
R=bradnelson@chromium.org, dpranke@chromium.org

Review URL: https://codereview.chromium.org/1431673005 .