libpthread: Move rdlock_count to be a thread-local variable
rdlock_count does not need to be defined in the nc_thread_descriptor
struct.  Since a thread's rdlock_count is only accessed by that thread
(not by other threads), rdlock_count can be a TLS variable.

This simplifies the code a little, and it gives a tiny memory saving
if nc_rwlock.c is not used.

This will be useful for PNaCl dynamic linking, because it reduces the
scope where nc_get_tdb() is used, making it easier to override its
definition for the dynamic-linking case.

BUG=https://bugs.chromium.org/p/nativeclient/issues/detail?id=4351
TEST=run_rwlock_test

Review URL: https://codereview.chromium.org/1817173002