[MIPS] Add support to signal_handler_regs_single_step test
Add arch-specific code to the signal_handler_regs_single_step test.
It mostly follows ARM version of the test, the exception is that removing
breakpoints has to take into account delay slots, i.e. breakpoints are
removed and restored for instructions in delay slots at the same time when
that happens for branches/jumps (that are followed by those delay slots).

BUG= No support for signal_handler_regs_single_step test for MIPS
TEST= ./scons platform=mips32 run_signal_handler_regs_single_step_test

Review-Url: https://codereview.chromium.org/2225153004