Set bot names back to precise for trybots
Infra is updating bots from Precise to Trusty and renaming them from 'precise' to 'linux'. The buildbots are finished but not the trybots. Rev 9f8fdd668b305511b605ee229533639c29a5222b updated the buildbot names but also some of the trybot names, so hack them back to 'precise' for now.

R=bradnelson@google.com, bradnelson
BUG= https://bugs.chromium.org/p/nativeclient/issues/detail?id=4388

Review-Url: https://codereview.chromium.org/2624913002 .